Added all validator tests and reworked all validators

+++ StringManager/StringManager.Services/API/Domain/Requests/ModifyToneRequest.cs
@@ -8,8 +8,8 @@ namespace StringManager.Services.API.Domain.Requests
 
         public string Name { get; set; }
 
-        public int? Frequency { get; set; }
+        public double? Frequency { get; set; }
 
-        public int? WaveLenght { get; set; }
+        public double? WaveLenght { get; set; }
     }
 }
